requirements.txt
numpy==1.21.0
pillow==9.0.1
requests==2.25.0
opencv-python-headless